projekty/code_twogether/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
requests
python-dotenv
flask_socketio
eventlet
